feat: combined datasets & eval - Evaluation updates - Train combined datasets ncdb-cls

+++ packnet_sfm/datasets/kitti_dataset_optimized.py
@@ -48,7 +48,7 @@ def read_png_depth(file):
     """Reads a .png depth map."""
     depth_png = np.array(load_image(file), dtype=int)
     assert (np.max(depth_png) > 255), 'Wrong .png depth file'
-    depth = depth_png.astype(np.float) / 256.
+    depth = depth_png.astype(np.float32) / 256.
     depth[depth_png == 0] = -1.
     return np.expand_dims(depth, axis=2)
